+++ src/Database/Create/Model.php
@@ -4,6 +4,13 @@ namespace Solital\Core\Database\Create;
 
 use Solital\Core\Database\ORM;
 
+/**
+ * This Model offers a way to access Katrina ORM's attributes and 
+ * classes without having to instantiate Katrina. 
+ * 
+ * This is an abstract class, so it should only be extended by other models. 
+ */
+
 abstract class Model
 {
     /**