+++ src/Resource/Helpers/helpers-http.php
@@ -11,7 +11,9 @@ use Solital\Core\Course\Course as Course;
  * @param string|null $name
  * @param string|array|null $parameters
  * @param array|null $getParams
+ * 
  * @return \Solital\Http\Uri
+ * 
  * @throws \InvalidArgumentException
  */
 function url(?string $name = null, $parameters = null, ?array $getParams = null): Uri
@@ -40,9 +42,10 @@ function request(): Request
  * @param string|null $index Parameter index name
  * @param string|null $defaultValue Default return value
  * @param array ...$methods Default methods
- * @return \Solital\Http\Input\InputHandler|array|string|null
+ * 
+ * @return mixed
  */
-function input($index = null, $defaultValue = null, ...$methods)
+function input(string $index = null, string $defaultValue = null, ...$methods)
 {
     if ($index !== null) {
         return request()->getInputHandler()->value($index, $defaultValue, ...$methods);
@@ -112,4 +115,52 @@ function redirect(string $url, ?int $code = null): void
 
     response()->redirect($url);
     exit;
+}
+
+/**
+ * @param string $key
+ * @param int $limit = 5
+ * @param int $seconds = 60
+ * 
+ * @return bool
+ */
+function request_limit(string $key, int $limit = 5, int $seconds = 60) : bool
+{
+    if (Session::has($key) && $_SESSION[$key]->time >= time() && $_SESSION[$key]->requests < $limit) {
+        Session::new($key, [
+            'time' => time() + $seconds,
+            'requests' => $_SESSION[$key]->requests + 1
+        ]);
+
+        return false;
+    }
+
+    if (Session::has($key) && $_SESSION[$key]->time >= time() && $_SESSION[$key]->requests >= $limit) {
+        return true;
+    }
+
+    Session::new($key, [
+        'time' => time() + $seconds,
+        'requests' => 1
+    ]);
+
+    return false;
+}
+
+/**
+ * request_repeat
+ *
+ * @param string $key
+ * @param string $value
+ * 
+ * @return bool
+ */
+function request_repeat(string $key, string $value): bool
+{
+    if (Session::has($key) && Session::show($key) == $value) {
+        return true;
+    }
+
+    Session::new($key, $value);
+    return false;
 }